requirements.txt
paho-mqtt
python-xsense
aiohttp